[DEV] iteration 3+4 DONE excluding AI skip ai: select, move, attack move range, attack range + visual feedback

+++ Assets/Scripts/Character/Data/CharacterData.cs
@@ -9,8 +9,8 @@ public class CharacterData
     public ECharacterType CharacterType;
     public int HealthPoints;
     public int DamagePoints;
-    public int MoveRange;
-    public int AttackRange;
+    public float MoveRange;
+    public float AttackRange;
     public GameObject Prefab;
 
 }
